use bigint/bigserial

+++ backend/setup.sql
@@ -6,12 +6,12 @@ GRANT ALL PRIVILEGES ON DATABASE foxcaves TO foxcaves;
 \c foxcaves foxcaves;
 
 CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     username VARCHAR(255),
     email VARCHAR(255),
     password VARCHAR(255),
     active INT NOT NULL DEFAULT 0,
-    bonusbytes INT NOT NULL DEFAULT 0,
+    bonusbytes BIGINT NOT NULL DEFAULT 0,
     loginkey VARCHAR(255),
     apikey VARCHAR(255)
 );
@@ -20,20 +20,20 @@ CREATE UNIQUE INDEX ON users (lower(email));
 
 CREATE TABLE files (
     id VARCHAR(32) PRIMARY KEY,
-    "user" INT REFERENCES users (id),
+    "user" BIGINT REFERENCES users (id),
     name VARCHAR(255),
     extension VARCHAR(255),
     type INT,
-    size INT,
-    time INT,
+    size BIGINT,
+    time BIGINT,
     thumbnail VARCHAR(255)
 );
 CREATE INDEX ON files ("user");
 
 CREATE TABLE links (
     id VARCHAR(32) PRIMARY KEY,
-    "user" INT REFERENCES users (id),
+    "user" BIGINT REFERENCES users (id),
     url VARCHAR(4096),
-    time INT
+    time BIGINT
 );
 CREATE INDEX ON links ("user");
